--- docs/source/geomapi/geomapi.nodes.geometrynode.rst
@@ -1,7 +1,0 @@
-geomapi.nodes.geometrynode
-==========================
-
-.. automodule:: geomapi.nodes.geometrynode
-   :members:
-   :show-inheritance:
-   :undoc-members:
