--- docs/source/geomapi/geomapi.nodes.node_refactor.rst
@@ -1,7 +1,0 @@
-geomapi.nodes.node\_refactor
-============================
-
-.. automodule:: geomapi.nodes.node_refactor
-   :members:
-   :show-inheritance:
-   :undoc-members:
