+++ docs/source/geomapi/geomapi.nodes.rst
@@ -6,16 +6,13 @@ geomapi.nodes
    :maxdepth: 4
 
    geomapi.nodes.bimnode
-   geomapi.nodes.geometrynode
    geomapi.nodes.imagenode
    geomapi.nodes.linesetnode
    geomapi.nodes.meshnode
    geomapi.nodes.node
-   geomapi.nodes.node_refactor
    geomapi.nodes.orthonode
    geomapi.nodes.panonode
    geomapi.nodes.pointcloudnode
-   geomapi.nodes.sessionnode_old
    geomapi.nodes.setnode
