--- docs/source/geomapi/geomapi.nodes.sessionnode_old.rst
@@ -1,7 +1,0 @@
-geomapi.nodes.sessionnode\_old
-==============================
-
-.. automodule:: geomapi.nodes.sessionnode_old
-   :members:
-   :show-inheritance:
-   :undoc-members:
