+++ docs/source/geomapi/geomapi.tools.validationtools.rst
@@ -2,12 +2,6 @@ geomapi.tools.validationtools
 =============================
 
 
-.. toctree::
-   :maxdepth: 4
-
-   geomapi.tools.validationtools.geomapi_updates
-
-
 .. automodule:: geomapi.tools.validationtools
    :members:
    :show-inheritance:
