+++ tests/test_tools.py
@@ -92,11 +92,9 @@ class TestTools(unittest.TestCase):
         #0
         self.assertAlmostEqual(nodes[0].cartesianTransform[0,3],0.379,delta=0.01)
         self.assertLess(nodes[0].cartesianTransform[0,0],0)
-        self.assertAlmostEqual(nodes[0].cartesianBounds[0],-4.351,delta=0.01)
         #1
         self.assertEqual(nodes[1].cartesianTransform[0,3],0.0)
         self.assertEqual(nodes[1].cartesianTransform[0,0],-1)
-        self.assertAlmostEqual(nodes[1].cartesianBounds[0],-5.69,delta=0.01)
         
         #with kwargs
         nodes=tl.e57xml_to_pointcloud_nodes(xmlPath=self.dataLoaderParking.e57XmlPath,myattrib=1)
@@ -141,13 +139,12 @@ class TestTools(unittest.TestCase):
         tl.e57_to_pointcloud_nodes(e57Path=self.dataLoaderParking.e57Path2)
         self.assertAlmostEqual(nodes[0].cartesianTransform[0,3],0.379,delta=0.01)
         self.assertLess(nodes[0].cartesianTransform[0,0],0)
-        self.assertAlmostEqual(nodes[0].cartesianBounds[0],-4.351,delta=0.01)
 
     def test_dxf_to_lineset_nodes(self):
         tl.dxf_to_lineset_nodes(dxfPath=self.dataLoaderRailway.dxfPath)
 
     def test_dxf_to_ortho_nodes(self):
-        nodes = tl.dxf_to_ortho_nodes(dxfPath=self.dataLoaderRailway.orthoDxfPath2)
+        nodes = tl.dxf_to_ortho_nodes(dxfPath=self.dataLoaderRailway.orthoDxfPath2,height=self.dataLoaderRailway.orthoHeight)
         #dxfPath and name + height-> offset in y and z
         node=nodes[0] #OrthoNode(dxfPath=self.dataLoaderRailway.orthoDxfPath2,name='railway-0-0',height=self.dataLoaderRailway.orthoHeight)
         self.assertEqual(node.dxfPath,self.dataLoaderRailway.orthoDxfPath2)
@@ -166,7 +163,7 @@ class TestTools(unittest.TestCase):
 
     def test_navvis_csv_to_pano_nodes(self):
         nodes=tl.navvis_csv_to_pano_nodes(csvPath =self.dataLoaderIndoorSite.csvPath)
-        self.assertEqual(len(nodes), 125)
+        self.assertEqual(len(nodes), 30)
 
     def select_nodes_k_nearest_neighbors(self):
         nodes=tl.graph_to_nodes(graphPath=self.dataLoaderRoad.ifcGraphPath,loadResource=True)
@@ -199,7 +196,7 @@ class TestTools(unittest.TestCase):
         nodes=tl.graph_to_nodes(graphPath=self.dataLoaderRoad.ifcGraphPath,loadResource=True)
         hull = nodes[0].convexHull
         list=tl.select_nodes_within_convex_hull(nodes,hull)
-        self.assertEqual(len(list),31)
+        self.assertEqual(len(list),1)
 
     def test_select_nodes_intersecting_bounding_box(self): 
         nodes=tl.graph_to_nodes(graphPath=self.dataLoaderRoad.ifcGraphPath,loadResource=True)
@@ -220,8 +217,6 @@ class TestTools(unittest.TestCase):
         subjects=[s for s in graph.subjects(RDF.type)]
         self.assertEqual(len(nodes), len(subjects))
         
-    def test_navvis_decode_depthmap(self):
-        tl.navvis_decode_depthmap(np.asarray(self.dataLoaderIndoorSite.depthImage))
 
 if __name__ == '__main__':
     unittest.main()
